+++ app/src/androidTest/java/woowacourse/movie/view/home/HomeFragmentTest.kt
@@ -17,7 +17,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import woowacourse.movie.MainActivity
+import woowacourse.movie.view.MainActivity
 import woowacourse.movie.R
 import woowacourse.movie.utils.MovieUtils.navigateToBottomMenu
 import woowacourse.movie.view.home.adapter.viewholder.AdvertisementViewHolder
