+++ app/src/main/java/woowacourse/movie/db/theater/TheaterDatabase.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.db.theater
 
-import woowacourse.movie.model.Theater
+import woowacourse.movie.model.theater.Theater
 import woowacourse.movie.model.movie.ScreeningTimes
 import java.time.LocalTime
