+++ app/src/main/java/woowacourse/movie/model/theater/Theater.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.theater
 
 import woowacourse.movie.model.movie.ScreeningTimes
