+++ app/src/main/java/woowacourse/movie/presenter/reservation/ReservationDetailContract.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.presenter.reservation
 
-import woowacourse.movie.model.HeadCount
+import woowacourse.movie.model.ticket.HeadCount
 import woowacourse.movie.model.movie.Movie
 import woowacourse.movie.model.movie.ScreeningDateTime
 import woowacourse.movie.model.movie.ScreeningTimes
