+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationDetailActivity.kt
@@ -12,7 +12,7 @@ import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivityReservationDetailBinding
 import woowacourse.movie.db.screening.ScreeningDao
 import woowacourse.movie.db.theater.TheaterDao
-import woowacourse.movie.model.HeadCount
+import woowacourse.movie.model.ticket.HeadCount
 import woowacourse.movie.model.movie.Movie
 import woowacourse.movie.model.movie.ScreeningDateTime
 import woowacourse.movie.model.movie.ScreeningTimes
