+++ app/src/main/java/woowacourse/movie/view/reservation/SeatSelectionActivity.kt
@@ -12,7 +12,7 @@ import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivitySeatSelectionBinding
 import woowacourse.movie.db.screening.ScreeningDao
 import woowacourse.movie.db.seats.SeatsDao
-import woowacourse.movie.model.HeadCount
+import woowacourse.movie.model.ticket.HeadCount
 import woowacourse.movie.model.movie.Movie
 import woowacourse.movie.model.movie.ScreeningDateTime
 import woowacourse.movie.model.seats.Grade
