+++ app/src/main/java/woowacourse/movie/view/theater/adapter/TheaterSelectionAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.databinding.ItemTheaterBinding
-import woowacourse.movie.model.Theater
+import woowacourse.movie.model.theater.Theater
 
 typealias OnSelectTheater = (theaterId: Int) -> Unit
