+++ app/src/main/res/layout/fragment_reservation_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ReservationHistoryFragment">
+    tools:context=".view.history.ReservationHistoryFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
