+++ app/src/main/res/layout/fragment_setting.xml
@@ -4,7 +4,7 @@
     android:background="@color/white"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SettingFragment">
+    tools:context=".view.setting.SettingFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
